+++ app-public-api/src/main/java/co/kurrant/app/public_api/service/FoodService.java
@@ -3,6 +3,7 @@ package co.kurrant.app.public_api.service;
 import co.dalicious.domain.food.dto.FoodDetailDto;
 import co.kurrant.app.public_api.dto.food.DailyFoodDto;
 
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.util.Date;
 import java.util.List;
@@ -11,5 +12,5 @@ public interface FoodService {
 
     Object getDailyFood(Integer spotId, LocalDate selectedDate);
 
-    FoodDetailDto getFoodDetail(Integer foodId);
+    FoodDetailDto getFoodDetail(BigInteger foodId);
 }
